ARROW-3382: [Gandiva][C++] Run tests in travis
Author: Pindikura Ravindra <ravindra@dremio.com>
Author: Wes McKinney <wesm+git@apache.org>

Closes #2705 from pravindra/arrow-3382 and squashes the following commits:

a312b3c92 <Pindikura Ravindra> ARROW-3382:  minor change in comment
77097deaf <Pindikura Ravindra> ARROW-3382:  run integ tests serially
5e9424e51 <Pindikura Ravindra> ARROW-3382:  fix a typo in travis.yml
b011c6445 <Pindikura Ravindra> ARROW-3382:  Set gcc to 4.9
d7da6107b <Pindikura Ravindra> ARROW-3382:  set C++ to g++4.9
512b7e0cf <Pindikura Ravindra> ARROW-3382:  use gcc 4.9
c80b8c62a <Pindikura Ravindra> ARROW-3382:  add dep for precompiled
714502dc7 <Pindikura Ravindra> ARROW-3382:  switch to static RE2
c7ad2a512 <Pindikura Ravindra> ARROW-3382:  fix compiler warnings
c594e647d <Pindikura Ravindra> ARROW-3382:  fix format error
8f881d5f0 <Pindikura Ravindra> ARROW-3382: fix compiler warnings
3001bc3c7 <Wes McKinney> Add option to dynamically-link re2, use in Gandiva build temporarily
fccf41043 <Pindikura Ravindra> ARROW-3382:  temp workaround re2
28fb3f314 <Pindikura Ravindra> ARROW-3382:  Run tests in travis

+++ cpp/src/gandiva/precompiled/hash_test.cc
@@ -44,8 +44,8 @@ TEST(TestHash, TestHash32) {
   EXPECT_EQ(hash32(u16, 0), zero_hash);
   EXPECT_EQ(hash32(s32, 0), zero_hash);
   EXPECT_EQ(hash32(u32, 0), zero_hash);
-  EXPECT_EQ(hash32(s64, 0), zero_hash);
-  EXPECT_EQ(hash32(u64, 0), zero_hash);
+  EXPECT_EQ(hash32(static_cast<double>(s64), 0), zero_hash);
+  EXPECT_EQ(hash32(static_cast<double>(u64), 0), zero_hash);
   EXPECT_EQ(hash32(f32, 0), zero_hash);
   EXPECT_EQ(hash32(f64, 0), zero_hash);
 
@@ -82,8 +82,8 @@ TEST(TestHash, TestHash64) {
   EXPECT_EQ(hash64(u16, 0), zero_hash);
   EXPECT_EQ(hash64(s32, 0), zero_hash);
   EXPECT_EQ(hash64(u32, 0), zero_hash);
-  EXPECT_EQ(hash64(s64, 0), zero_hash);
-  EXPECT_EQ(hash64(u64, 0), zero_hash);
+  EXPECT_EQ(hash64(static_cast<double>(s64), 0), zero_hash);
+  EXPECT_EQ(hash64(static_cast<double>(u64), 0), zero_hash);
   EXPECT_EQ(hash64(f32, 0), zero_hash);
   EXPECT_EQ(hash64(f64, 0), zero_hash);
